+++ src/p_inter.c
@@ -1334,8 +1334,8 @@ void P_DamageMobj(mobj_t *target, mobj_t *inflicter, mobj_t *source, int damage,
     // Some close combat weapons should not
     // inflict thrust and push the victim out of reach,
     // thus kick away unless using the chainsaw.
-    if (inflicter && !(flags & MF_NOCLIP) && (!source || !splayer
-        || splayer->readyweapon != wp_chainsaw))
+    if (inflicter && (!tplayer || !inflicter->player) && !(flags & MF_NOCLIP)
+        && (!source || !splayer || splayer->readyweapon != wp_chainsaw))
     {
         unsigned int    ang = R_PointToAngle2(inflicter->x, inflicter->y, target->x, target->y);
         int             mass = (corpse ? MAX(200, info->mass) : info->mass);
@@ -1400,7 +1400,10 @@ void P_DamageMobj(mobj_t *target, mobj_t *inflicter, mobj_t *source, int damage,
 
     if (tplayer)
     {
-        int     damagecount;
+        int damagecount;
+
+        if (freeze && (!inflicter || !inflicter->player))
+            return;
 
         // end of game hell hack
         if (target->subsector->sector->special == DamageNegative10Or20PercentHealthAndEndLevel
@@ -1409,7 +1412,7 @@ void P_DamageMobj(mobj_t *target, mobj_t *inflicter, mobj_t *source, int damage,
 
         // Below certain threshold,
         // ignore damage in GOD mode, or with INVUL power.
-        if ((tplayer->cheats & CF_GODMODE) || freeze || idclevtics
+        if ((tplayer->cheats & CF_GODMODE) || idclevtics
             || (damage < 1000 && tplayer->powers[pw_invulnerability]))
             return;