+++ docs/releasenotes.md
@@ -27,6 +27,7 @@
 * When exiting a map that has no monsters to kill and/or no items to pick up, the intermission screen will now show `100%` kills and/or items instead of `0%`.
 * The map name will now be displayed on the intermission screen before starting MAP31 in *DOOM II* and MAP09 in *DOOM II: No Rest For The Living*.
 * Monsters are now gibbed rather than just becoming a pool of blood splats when crushed by a lowering platform or door.
+* The `health` CVAR can now be reduced when freeze mode is on.
 
 ---